Create FeatureSet.py
This implements the code to *generate* a set of natural classes from a feature set.

Semi-formally, a *featurization* is defined as a tuple (SIGMA, FEATURE_NAMES, FEATURES) where SIGMA is the segmental alphabet, FEATURE_NAMES is a list of names for features, and FEATURES : FEATURE_NAMES x SIGMA --> {-, +, 0}, i.e. each feature is a predicate mapping segments to feature values.